# docs/03_design/openapi.yaml
openapi: 3.0.3
info:
  title: 사내 연차 경매 시스템 API
  description: |
    B2E Escrow & Dividend 모델 기반 연차 경매 플랫폼의 정식 REST API 명세.

    - 인증: Bearer JWT (SSO 발급)
    - 모든 응답은 `ApiSuccess` 또는 `ApiError` 봉투(envelope)로 감싸짐
    - 내부 호출 엔드포인트(`/internal/*`)는 `X-Internal-Token` 헤더 필수 — 외부 노출 금지
    - 권한 검증은 `permission-matrix.md` 1:1 매핑
  version: 0.1.0
  contact:
    name: 타임소프트콘 (김기철, 오지석)
servers:
  - url: https://auction.company.internal/api/v1
    description: 운영 (예정)
  - url: http://localhost:3000/api/v1
    description: 로컬 개발
...
security:
  - bearerAuth: []

# ============================================================================
# Paths
# ============================================================================
paths:
...
  /auctions:
    get:
      tags: [Auction]
      summary: 경매 목록 조회 (FR-2.1)
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/AuctionStatus'
        - $ref: '#/components/parameters/YearQuery'
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/SizeQuery'
      responses:
        '200':
          description: 경매 목록
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/PaginatedAuctions' }
  /auctions/{id}:
    get:
      tags: [Auction]
      summary: 경매 상세 조회
      parameters:
        - $ref: '#/components/parameters/AuctionIdPath'
      responses:
        '200':
          description: 경매 상세
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/AuctionDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
  /auctions/{id}/bids:
    post:
      tags: [Auction]
      summary: 입찰 (FR-2.1)
      description: |
        SQLite write 락(lockAuction) → wallet 차감 → 직전자 즉시 환불 → highest_bid 갱신 → LEDGER_ENTRY → WebSocket 브로드캐스트. 모두 단일 트랜잭션. scope-cuts CUT-1(ADR-006 Superseded)·ADR-018 참조.
      parameters:
        - $ref: '#/components/parameters/AuctionIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BidRequest' }
      responses:
        '200':
          description: 입찰 수락
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/BidResult' }
        '400':
          description: |
            - `POINT_INSUFFICIENT` — wallet 잔액 부족
            - `BID_TOO_LOW` — 현재가 + 최소 증분 미만
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '409':
          description: |
            - `AUCTION_CLOSED` — 마감된 경매
            - `INVALID_STATE_TRANSITION` — OPEN 상태가 아님
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
  parameters:
    AuctionIdPath:
      in: path
      name: id
      required: true
      schema: { type: integer, format: int64 }
    YearQuery:
      in: query
      name: year
      schema: { type: integer, example: 2026 }
...
    PageQuery:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    SizeQuery:
      in: query
      name: size
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
...
    NotFound:
      description: 리소스 없음
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
...
    ApiSuccess:
      type: object
      required: [success, data, timestamp]
      properties:
        success: { type: boolean, enum: [true] }
        data: { type: object, nullable: true }
        error: { type: object, nullable: true, example: null }
        timestamp: { type: string, format: date-time }
    ApiError:
      type: object
      required: [success, error, timestamp]
      properties:
        success: { type: boolean, enum: [false] }
        data: { type: object, nullable: true, example: null }
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: |
                표준 에러 코드:
                - UNAUTHORIZED · FORBIDDEN · NOT_FOUND
                - POINT_INSUFFICIENT · BID_TOO_LOW · INVALID_STATE_TRANSITION
                - AUCTION_CLOSED
                - REASON_REQUIRED · INVALID_AMOUNT · USER_NOT_FOUND
                - HR_API_TIMEOUT · ESCROW_MISMATCH
                - INSUFFICIENT_LEAVE
            message: { type: string }
            details: { type: object, nullable: true }
        timestamp: { type: string, format: date-time }

    # ---- 열거형 ----
    AuctionStatus:
      type: string
      enum: [CREATED, OPEN, CLOSED, AWARDED, UNSOLD, EXPIRED]
      description: Auction 6상태 (ADR-014 State 패턴)
...
    AuctionSummary:
      type: object
      properties:
        id: { type: integer, format: int64 }
        status: { $ref: '#/components/schemas/AuctionStatus' }
        year: { type: integer }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        highestBid: { type: integer, format: int64 }
        myCurrentBid:
          type: integer
          format: int64
          nullable: true
          description: 내가 현재 최고가 입찰자인 경우의 입찰액
        timeRemainingSec: { type: integer }
    AuctionDetail:
      allOf:
        - $ref: '#/components/schemas/AuctionSummary'
        - type: object
          properties:
            winnerId: { type: integer, format: int64, nullable: true }
            startingPrice: { type: integer, format: int64 }
            minBidIncrement: { type: integer, format: int64 }
    PaginatedAuctions:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/AuctionSummary' }
        total: { type: integer }
        page: { type: integer }
        size: { type: integer }
    BidRequest:
      type: object
      required: [amount]
      properties:
        amount:
          type: integer
          format: int64
          minimum: 1
          description: 입찰 금액 (포인트)
    BidResult:
      type: object
      properties:
        auctionId: { type: integer, format: int64 }
        acceptedAmount: { type: integer, format: int64 }
        newHighestBid: { type: integer, format: int64 }
        refundedPreviousBidder:
          type: integer
          format: int64
          nullable: true
          description: 즉시 환불된 직전 최고가 입찰자 ID (ADR-018)
...